requirements.txt
adjustText==1.3.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.15.8
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
BiocFrame==0.5.11
biocutils==0.1.7
biomart==0.9.2
bleach==6.2.0
bokeh==3.6.1
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2024.11.0
dask-jobqueue==0.9.0
dask_labextension==7.0.0
datacache==1.4.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distributed==2024.11.0
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.1
forestci==0.7
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
GenomicRanges==0.4.34
git-filter-repo==2.47.0
greenlet==3.1.1
grpcio==1.68.0
gtfparse==2.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
icecream==2.1.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
interlap==0.2.7
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
IRanges==0.2.12
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
-e git+ssh://git@bitbucket.org/benccia/methylationclassifier.git@fa1fc1719d01ec77ffcd88d100b1d0998133dc4f#egg=mch
memoized-property==1.0.3
methylcheck==0.8.5
methylize==1.1.1
methylprep==1.7.1
mistune==3.0.2
mock==5.1.0
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ncls==0.0.68
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parse==1.18.0
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polars==1.12.0
progressbar33==2.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyensembl==2.3.13
Pygments==2.18.0
pylint==2.17.7
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
result==0.17.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
rpy2==3.5.16
rpy2-arrow==0.1.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
serializable==0.4.1
simpervisor==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlmodel==0.0.22
stack-data==0.6.3
statsmodels==0.14.4
tblib==3.0.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tinytimer==0.0.0
tomlkit==0.13.2
toolshed==0.4.6
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typechecks==0.1.0
typedb-client==2.18.2
typedb-protocol==2.18.0
types-python-dateutil==2.9.0.20241003
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.0
XlsxWriter==3.2.3
xyzservices==2024.9.0
yarl==1.17.1
zict==3.0.0
zipp==3.21.0
